// src/enums/node-caption.ts
export enum NodeCaption
{
    exported = "exported",
    notExported = "not exported",
    private = "private",
    protected = "protected",
    public = "public",
    enum = "enum",
    interface = "interface",
    class = "class",
    type = "type",
    constVariable = "const variable",
    variable = "variable",
    function = "function",
    enumValue = "enum values",
    readonlyProperty = "readonly property",
    property = "property",
    constructor = "constructor",
    staticConstructor = "static constructor",
    index = "index",
    accessor = "accessor",
    getter = "getter",
    setter = "setter",
    method = "method",
}